if client revokes current app id, log them out
