+++ src/templates/user/security/authorised_apps.html.tt
@@ -2,6 +2,7 @@
     <div class="page-title" id="applications-title">
         <h1>[% l("Applications") %]</h1>
         <p id="description">[% l("Keep track of your authorised applications.") %]</p>
-        <p class="error-msg" id="err"></p>
+        <p class="error-msg" id="applications_error"></p>
     </div>
+    <div id="applications_loading">[% PROCESS global/loading.html.tt %]</div>
 </div>
